refactor: new sub-reconciler for monitoring federation
